mongodb_cmdline_tool/requirements.txt
jira ~= 1.0
keyring ~= 12.0
invoke ~= 1.0
pyyaml ~= 3.0
